Conflicts:
	xnradmin-model/xnradmin-core/src/main/java/com/xnradmin/core/action/stat/OrderRecordStatAction.java
	xnradmin-model/xnradmin-data/src/main/java/com/xnradmin/po/business/BusinessOrderGoodsRelation.java
	xnradmin-web/src/main/webapp/stat/order/modify.jsp

+++ xnradmin-model/xnradmin-core/src/main/java/com/xnradmin/core/dao/business/order/AllocationListDAO.java
@@ -53,7 +53,7 @@ public class AllocationListDAO{
         }
     }
 
-    public AllocationData findById(Long id){
+    public AllocationData findById(Integer id){
         log.debug("getting AllocationData instance with id: " + id);
         try{
